+++ php/async/gate.php
@@ -57,10 +57,12 @@ class gate extends Exchange {
                     'public' => array(
                         'futures' => 'https://fx-api-testnet.gateio.ws/api/v4',
                         'delivery' => 'https://fx-api-testnet.gateio.ws/api/v4',
+                        'options' => 'https://fx-api-testnet.gateio.ws/api/v4',
                     ),
                     'private' => array(
                         'futures' => 'https://fx-api-testnet.gateio.ws/api/v4',
                         'delivery' => 'https://fx-api-testnet.gateio.ws/api/v4',
+                        'options' => 'https://fx-api-testnet.gateio.ws/api/v4',
                     ),
                 ),
                 'referral' => array(
@@ -444,6 +446,7 @@ class gate extends Exchange {
                 'X-Gate-Channel-Id' => 'ccxt',
             ),
             'options' => array(
+                'sandboxMode' => false,
                 'createOrder' => array(
                     'expiration' => 86400, // for conditional orders
                 ),
@@ -679,6 +682,11 @@ class gate extends Exchange {
         ));
     }
 
+    public function set_sandbox_mode($enable) {
+        parent::set_sandbox_mode($enable);
+        $this->options['sandboxMode'] = $enable;
+    }
+
     public function fetch_markets($params = array ()) {
         return Async\async(function () use ($params) {
             /**
@@ -686,15 +694,20 @@ class gate extends Exchange {
              * @param {array} $params extra parameters specific to the exchange api endpoint
              * @return {[array]} an array of objects representing market data
              */
+            $sandboxMode = $this->safe_value($this->options, 'sandboxMode', false);
             $promises = array(
-                $this->fetch_spot_markets($params),
                 $this->fetch_contract_markets($params),
                 $this->fetch_option_markets($params),
             );
+            if (!$sandboxMode) {
+                // gate does not have a sandbox for spot $markets
+                $mainnetOnly = array( $this->fetch_spot_markets($params) );
+                $promises = $this->array_concat($promises, $mainnetOnly);
+            }
             $promises = Async\await(Promise\all($promises));
-            $spotMarkets = $promises[0];
-            $contractMarkets = $promises[1];
-            $optionMarkets = $promises[2];
+            $spotMarkets = $this->safe_value($promises, 0, array());
+            $contractMarkets = $this->safe_value($promises, 1, array());
+            $optionMarkets = $this->safe_value($promises, 2, array());
             $markets = $this->array_concat($spotMarkets, $contractMarkets);
             return $this->array_concat($markets, $optionMarkets);
         }) ();