+++ php/async/okx.php
@@ -253,6 +253,7 @@ class okx extends Exchange {
                         'trade/fills-history' => 2,
                         'trade/orders-algo-pending' => 1,
                         'trade/orders-algo-history' => 1,
+                        'trade/order-algo' => 1,
                         'account/subaccount/balances' => 10,
                         'asset/subaccount/bills' => 5 / 3,
                         'users/subaccount/list' => 10,
@@ -2649,7 +2650,12 @@ class okx extends Exchange {
             $method = $this->safe_string($params, 'method', $defaultMethod);
             $stop = $this->safe_value($params, 'stop');
             if ($stop) {
-                throw new NotSupported($this->id . ' fetchOrder() does not support $stop orders, use fetchOpenOrders() fetchCanceledOrders() or fetchClosedOrders()');
+                $method = 'privateGetTradeOrderAlgo';
+                if ($clientOrderId !== null) {
+                    $request['algoClOrdId'] = $clientOrderId;
+                } else {
+                    $request['algoId'] = $id;
+                }
             } else {
                 if ($clientOrderId !== null) {
                     $request['clOrdId'] = $clientOrderId;
@@ -2657,7 +2663,7 @@ class okx extends Exchange {
                     $request['ordId'] = $id;
                 }
             }
-            $query = $this->omit($params, array( 'method', 'clOrdId', 'clientOrderId' ));
+            $query = $this->omit($params, array( 'method', 'clOrdId', 'clientOrderId', 'stop' ));
             $response = Async\await($this->$method (array_merge($request, $query)));
             //
             // Spot and Swap
@@ -2701,6 +2707,58 @@ class okx extends Exchange {
             //             }
             //         ),
             //         "msg" => ""
+            //     }
+            //
+            // Algo $order
+            //     {
+            //         "code":"0",
+            //         "msg":"",
+            //         "data":array(
+            //             {
+            //                 "instType":"FUTURES",
+            //                 "instId":"BTC-USD-200329",
+            //                 "ordId":"123445",
+            //                 "ccy":"BTC",
+            //                 "clOrdId":"",
+            //                 "algoId":"1234",
+            //                 "sz":"999",
+            //                 "closeFraction":"",
+            //                 "ordType":"oco",
+            //                 "side":"buy",
+            //                 "posSide":"long",
+            //                 "tdMode":"cross",
+            //                 "tgtCcy" => "",
+            //                 "state":"effective",
+            //                 "lever":"20",
+            //                 "tpTriggerPx":"",
+            //                 "tpTriggerPxType":"",
+            //                 "tpOrdPx":"",
+            //                 "slTriggerPx":"",
+            //                 "slTriggerPxType":"",
+            //                 "triggerPx":"99",
+            //                 "triggerPxType":"last",
+            //                 "ordPx":"12",
+            //                 "actualSz":"",
+            //                 "actualPx":"",
+            //                 "actualSide":"",
+            //                 "pxVar":"",
+            //                 "pxSpread":"",
+            //                 "pxLimit":"",
+            //                 "szLimit":"",
+            //                 "tag" => "adadadadad",
+            //                 "timeInterval":"",
+            //                 "callbackRatio":"",
+            //                 "callbackSpread":"",
+            //                 "activePx":"",
+            //                 "moveTriggerPx":"",
+            //                 "reduceOnly" => "false",
+            //                 "triggerTime":"1597026383085",
+            //                 "last" => "16012",
+            //                 "failCode" => "",
+            //                 "algoClOrdId" => "",
+            //                 "cTime":"1597026383000"
+            //             }
+            //         )
             //     }
             //
             $data = $this->safe_value($response, 'data', array());