+++ php/pro/ascendex.php
@@ -7,6 +7,7 @@ namespace ccxt\pro;
 
 use Exception; // a common import
 use ccxt\AuthenticationError;
+use ccxt\NetworkError;
 use React\Async;
 
 class ascendex extends \ccxt\async\ascendex {
@@ -492,6 +493,7 @@ class ascendex extends \ccxt\async\ascendex {
     public function watch_orders($symbol = null, $since = null, $limit = null, $params = array ()) {
         return Async\async(function () use ($symbol, $since, $limit, $params) {
             /**
+             * @see https://ascendex.github.io/ascendex-pro-api/#$channel-order-and-balance
              * watches information on multiple $orders made by the user
              * @param {string|null} $symbol unified $market $symbol of the $market $orders were made in
              * @param {int|null} $since the earliest time in ms to fetch $orders for
@@ -508,7 +510,7 @@ class ascendex extends \ccxt\async\ascendex {
             list($type, $query) = $this->handle_market_type_and_params('watchOrders', $market, $params);
             $messageHash = null;
             $channel = null;
-            if ($type !== 'spot') {
+            if ($type !== 'spot' && $type !== 'margin') {
                 $channel = 'futures-order';
                 $messageHash = 'order:FUTURES';
             } else {
@@ -947,14 +949,21 @@ class ascendex extends \ccxt\async\ascendex {
     }
 
     public function handle_ping($client, $message) {
-        $this->spawn(array($this, 'pong'), $client, $message);
+        return Async\async(function () use ($client, $message) {
+            try {
+                Async\await($this->spawn(array($this, 'pong'), $client, $message));
+            } catch (Exception $e) {
+                $error = new NetworkError ($this->id . ' handlePing failed with $error ' . $this->json($e));
+                $client->reset ($error);
+            }
+        }) ();
     }
 
     public function authenticate($url, $params = array ()) {
         $this->check_required_credentials();
         $messageHash = 'authenticated';
         $client = $this->client($url);
-        $future = $this->safe_value($client->futures, $messageHash);
+        $future = $this->safe_value($client->subscriptions, $messageHash);
         if ($future === null) {
             $timestamp = (string) $this->milliseconds();
             $urlParts = explode('/', $url);