+++ ts/src/pro/ascendex.ts
@@ -2,7 +2,7 @@
 //  ---------------------------------------------------------------------------
 
 import ascendexRest from '../ascendex.js';
-import { AuthenticationError } from '../base/errors.js';
+import { AuthenticationError, NetworkError } from '../base/errors.js';
 import { ArrayCache, ArrayCacheByTimestamp, ArrayCacheBySymbolById } from '../base/ws/Cache.js';
 //  ---------------------------------------------------------------------------
 
@@ -481,6 +481,7 @@ export default class ascendex extends ascendexRest {
         /**
          * @method
          * @name ascendex#watchOrders
+         * @see https://ascendex.github.io/ascendex-pro-api/#channel-order-and-balance
          * @description watches information on multiple orders made by the user
          * @param {string|undefined} symbol unified market symbol of the market orders were made in
          * @param {int|undefined} since the earliest time in ms to fetch orders for
@@ -497,7 +498,7 @@ export default class ascendex extends ascendexRest {
         const [ type, query ] = this.handleMarketTypeAndParams ('watchOrders', market, params);
         let messageHash = undefined;
         let channel = undefined;
-        if (type !== 'spot') {
+        if (type !== 'spot' && type !== 'margin') {
             channel = 'futures-order';
             messageHash = 'order:FUTURES';
         } else {
@@ -932,15 +933,20 @@ export default class ascendex extends ascendexRest {
         await client.send ({ 'op': 'pong', 'hp': this.safeInteger (message, 'hp') });
     }
 
-    handlePing (client, message) {
-        this.spawn (this.pong, client, message);
+    async handlePing (client, message) {
+        try {
+            await this.spawn (this.pong, client, message);
+        } catch (e) {
+            const error = new NetworkError (this.id + ' handlePing failed with error ' + this.json (e));
+            client.reset (error);
+        }
     }
 
     authenticate (url, params = {}) {
         this.checkRequiredCredentials ();
         const messageHash = 'authenticated';
         const client = this.client (url);
-        let future = this.safeValue (client.futures, messageHash);
+        let future = this.safeValue (client.subscriptions, messageHash);
         if (future === undefined) {
             const timestamp = this.milliseconds ().toString ();
             const urlParts = url.split ('/');