+++ ts/src/pro/bybit.ts
@@ -782,7 +782,8 @@ export default class bybit extends bybitRest {
             symbol = this.symbol (symbol);
             messageHash += ':' + symbol;
         }
-        const isUnifiedMargin = await this.isUnifiedMarginEnabled ();
+        const unified = await this.isUnifiedEnabled ();
+        const isUnifiedMargin = this.safeValue (unified, 0, false);
         const url = this.getUrlByMarketType (symbol, true, isUnifiedMargin, method, params);
         await this.authenticate (url);
         const topicByMarket = {
@@ -898,7 +899,8 @@ export default class bybit extends bybitRest {
             symbol = this.symbol (symbol);
             messageHash += ':' + symbol;
         }
-        const isUnifiedMargin = await this.isUnifiedMarginEnabled ();
+        const unified = await this.isUnifiedEnabled ();
+        const isUnifiedMargin = this.safeValue (unified, 0, false);
         const url = this.getUrlByMarketType (undefined, true, isUnifiedMargin, method, params);
         await this.authenticate (url);
         const topicsByMarket = {
@@ -1163,7 +1165,8 @@ export default class bybit extends bybitRest {
          */
         const method = 'watchBalance';
         const messageHash = 'balances';
-        const isUnifiedMargin = await this.isUnifiedMarginEnabled ();
+        const unified = await this.isUnifiedEnabled ();
+        const isUnifiedMargin = this.safeValue (unified, 0, false);
         const url = this.getUrlByMarketType (undefined, true, isUnifiedMargin, method, params);
         await this.authenticate (url);
         const topicByMarket = {