+++ ts/src/pro/whitebit.ts
@@ -794,7 +794,7 @@ export default class whitebit extends whitebitRest {
     async authenticate (params = {}) {
         this.checkRequiredCredentials ();
         const url = this.urls['api']['ws'];
-        const messageHash = 'login';
+        const messageHash = 'authenticated';
         const client = this.client (url);
         const future = client.future ('authenticated');
         const authenticated = this.safeValue (client.subscriptions, messageHash);
@@ -819,7 +819,12 @@ export default class whitebit extends whitebitRest {
                 'id': id,
                 'method': this.handleAuthenticate,
             };
-            this.spawn (this.watch, url, messageHash, request, messageHash, subscription);
+            try {
+                await this.watch (url, messageHash, request, messageHash, subscription);
+            } catch (e) {
+                delete client.subscriptions[messageHash];
+                future.reject (e);
+            }
         }
         return await future;
     }
@@ -851,8 +856,8 @@ export default class whitebit extends whitebitRest {
         } catch (e) {
             if (e instanceof AuthenticationError) {
                 client.reject (e, 'authenticated');
-                if ('login' in client.subscriptions) {
-                    delete client.subscriptions['login'];
+                if ('authenticated' in client.subscriptions) {
+                    delete client.subscriptions['authenticated'];
                 }
                 return false;
             }